+++ backend/farm/planning/__init__.py
@@ -1,0 +1,1 @@
+"""Planning domain: planting plans, tasks, and the yield calendar."""
